requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_ipyw_jlab_nb_ext_conf=0.1.0=py37_0
anaconda-client=1.7.2=py37_0
anaconda-navigator=1.9.12=py37_0
anyio=2.2.0=py37hecd8cb5_1
appnope=0.1.2=py37hecd8cb5_1001
argon2-cffi=20.1.0=py37h9ed2024_1
async_generator=1.10=py37h28b3542_0
attrs=21.2.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
backports=1.0=pyhd3eb1b0_2
backports.functools_lru_cache=1.6.4=pyhd3eb1b0_0
backports.tempfile=1.0=pyhd3eb1b0_1
backports.weakref=1.0.post1=py_1
beautifulsoup4=4.9.3=pyha847dfd_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
brotlipy=0.7.0=py37h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.5.25=hecd8cb5_1
cairo=1.16.0=h0ab9d94_1001
certifi=2021.5.30=py37hecd8cb5_0
cffi=1.14.0=py37hb5b8e2f_0
chardet=4.0.0=py37hecd8cb5_1003
click=8.0.1=pyhd3eb1b0_0
clyent=1.2.2=py37_1
conda=4.10.1=py37hecd8cb5_1
conda-build=3.18.11=py37_0
conda-env=2.6.0=1
conda-package-handling=1.7.3=py37h9ed2024_1
conda-verify=3.4.2=py_1
cryptography=3.4.7=py37h2fd3fbb_0
cycler=0.10.0=py_2
decorator=5.0.9=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py37_0
ffmpeg=4.0=h01ea3c9_0
filelock=3.0.12=pyhd3eb1b0_1
fontconfig=2.13.1=h10f422b_1005
freetype=2.10.4=ha233b18_0
future=0.18.2=py37_1
gettext=0.19.8.1=haf92f58_1004
glib=2.58.3=py37h7c187be_1004
glob2=0.7=pyhd3eb1b0_0
graphite2=1.3.13=h12caacf_1001
harfbuzz=1.8.8=hb8d4a28_0
hdf5=1.10.2=hc401514_3
icu=58.2=h0a44026_3
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.10.0=py37hecd8cb5_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2021.2.0=hecd8cb5_564
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.22.0=py37h01d92e1_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_1
jasper=2.0.14=h636a363_1
jedi=0.17.0=py37_0
jinja2=3.0.0=pyhd3eb1b0_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9b=he5867d9_2
json5=0.9.5=py_0
jsonschema=3.2.0=py_2
jupyter-packaging=0.7.12=pyhd3eb1b0_0
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py37hecd8cb5_0
jupyter_server=1.4.1=py37hecd8cb5_0
jupyterlab=3.0.14=pyhd3eb1b0_1
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.4.0=pyhd3eb1b0_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.1=py37h23ab428_0
lcms2=2.12=hf1fd2bf_0
libarchive=3.4.2=haa3ed63_0
libblas=3.9.0=9_mkl
libcblas=3.9.0=9_mkl
libcxx=10.0.0=1
libedit=3.1.20210216=h9ed2024_1
libffi=3.2.1=h0a44026_1007
libgfortran=3.0.1=0
libiconv=1.16=h1de35cc_0
liblapack=3.9.0=9_mkl
liblief=0.10.1=h0a44026_0
libopencv=3.4.2=h7c891bd_1
libopus=1.3.1=hc929b4f_1
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.1.0=hcb84e12_1
libvpx=1.7.0=h378b8a2_0
libxml2=2.9.10=h7cdb67c_3
llvm-openmp=11.1.0=hda6cdc1_1
lz4-c=1.9.3=h23ab428_0
markupsafe=2.0.1=py37h9ed2024_0
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py37hddda452_1
mistune=0.8.4=py37h1de35cc_0
mkl=2021.2.0=hecd8cb5_269
mkl-service=2.3.0=py37h9ed2024_1
mkl_fft=1.3.0=py37h4a7008c_2
mkl_random=1.2.1=py37hb2f4e1b_2
navigator-updater=0.2.1=py37_0
nbclassic=0.2.6=pyhd3eb1b0_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py37_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
ninja=1.10.2=hf7b0b51_1
notebook=6.4.0=py37hecd8cb5_0
numpy=1.20.2=py37h4b4dc7a_0
numpy-base=1.20.2=py37he0bd621_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
opencv-python=4.5.2.54=pypi_0
openssl=1.1.1k=h9ed2024_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.2.4=py37h23ab428_0
pandoc=2.12=hecd8cb5_0
pandocfilters=1.4.3=py37hecd8cb5_1
parso=0.8.2=pyhd3eb1b0_0
pcre=8.44=h4a8c4bd_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py37h5270095_0
pip=21.1.1=py37hecd8cb5_0
pixman=0.38.0=h01d97ff_1003
pkginfo=1.7.0=py37hecd8cb5_0
prometheus_client=0.10.1=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
psutil=5.8.0=py37h9ed2024_1
ptyprocess=0.7.0=pyhd3eb1b0_2
py-lief=0.10.1=py37haf313ee_0
py-opencv=3.4.2=py37h7c891bd_1
pycosat=0.6.3=py37h9ed2024_0
pycparser=2.20=py_2
pygments=2.9.0=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.6.0=py37h655552a_6
pyrsistent=0.17.3=py37haf1e3a3_0
pysocks=1.7.1=py37hecd8cb5_0
python=3.7.6=h359304d_2
python-dateutil=2.8.1=pyhd3eb1b0_0
python-libarchive-c=2.9=pyhd3eb1b0_1
python.app=2=py37_10
python_abi=3.7=1_cp37m
pytorch=1.5.0=py3.7_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=py37h9ed2024_1
pyzmq=20.0.0=py37h23ab428_1
qt=5.6.3=h1d42b2f_0
qtpy=1.9.0=py_0
readline=7.0=h1de35cc_5
requests=2.25.1=pyhd3eb1b0_0
ripgrep=12.1.1=0
ruamel_yaml=0.15.100=py37h9ed2024_0
scikit-learn=0.22.1=py37h3dc85bc_1
scipy=1.2.1=py37hbd7caa9_1
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py37hecd8cb5_0
sip=4.18.1=py37h0a44026_2
six=1.15.0=py37hecd8cb5_0
sniffio=1.2.0=py37hecd8cb5_1
soupsieve=2.2.1=pyhd3eb1b0_0
sqlite=3.33.0=hffcf06c_0
terminado=0.9.4=py37hecd8cb5_0
testpath=0.4.4=pyhd3eb1b0_0
tk=8.6.10=hb0a8c7a_0
torchvision=0.6.0=py37_cpu
tornado=6.1=py37h9ed2024_0
tqdm=4.59.0=pyhd3eb1b0_1
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.4=pyhd3eb1b0_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py37_0
xmltodict=0.12.0=py_0
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
zeromq=4.3.4=h23ab428_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
